[Refactor] Comment import
